requirements.txt
asgiref==3.7.2
Django==5.0
django-appconf==1.0.6
django-ckeditor==6.7.0
django-filter==23.5
django-imagekit==5.0.0
django-js-asset==2.2.0
django-modeladmin-reorder==0.3.1
flake8==6.1.0
isort==5.13.2
mccabe==0.7.0
pilkit==3.0
Pillow==10.1.0
pycodestyle==2.11.1
pyflakes==3.1.0
sqlparse==0.4.4
